+++ POS/src/utils/offline/workerClient.js
@@ -438,6 +438,10 @@ class OfflineWorkerClient {
 		return this.sendMessage("DELETE_CUSTOMERS", { customerNames })
 	}
 
+	async deleteItems(itemCodes) {
+		return this.sendMessage("DELETE_ITEMS", { itemCodes })
+	}
+
 	async cachePaymentMethods(paymentMethods) {
 		return this.sendMessage("CACHE_PAYMENT_METHODS", { paymentMethods })
 	}